PARQUET-456: Finish gzip implementation and unit test all compressors
We should perhaps separate compression and decompression code (as in Impala) as gzip is more stateful than the other compressors.

Closes #11 when merged.

Author: Wes McKinney <wes@cloudera.com>
Author: Konstantin Knizhnik <knizhnik@garret.ru>

Closes #48 from wesm/PARQUET-456 and squashes the following commits:

5aeba2a [Wes McKinney] Comment typo
8e1f8f2 [Wes McKinney] Move test run to shell script and enable OS X
633fd71 [Wes McKinney] Port gzip codec code from Impala, expand tests, get them to pass
a8d3c11 [Wes McKinney] Add compression round-trip test, gzip needs a bunch more work though
0bc8cf7 [Wes McKinney] Fix PATH_SUFFIXES for zlib
69548c9 [Konstantin Knizhnik] Add zlib to thirdparty build toolchain for compression codec

+++ cpp/src/parquet/compression/lz4-codec.cc
@@ -21,21 +21,21 @@
 
 namespace parquet_cpp {
 
-void Lz4Codec::Decompress(int input_len, const uint8_t* input,
-      int output_len, uint8_t* output_buffer) {
-  int n = LZ4_decompress_fast(reinterpret_cast<const char*>(input),
+void Lz4Codec::Decompress(int64_t input_len, const uint8_t* input,
+      int64_t output_len, uint8_t* output_buffer) {
+  int64_t n = LZ4_decompress_fast(reinterpret_cast<const char*>(input),
       reinterpret_cast<char*>(output_buffer), output_len);
   if (n != input_len) {
     throw parquet_cpp::ParquetException("Corrupt lz4 compressed data.");
   }
 }
 
-int Lz4Codec::MaxCompressedLen(int input_len, const uint8_t* input) {
+int64_t Lz4Codec::MaxCompressedLen(int64_t input_len, const uint8_t* input) {
   return LZ4_compressBound(input_len);
 }
 
-int Lz4Codec::Compress(int input_len, const uint8_t* input,
-    int output_buffer_len, uint8_t* output_buffer) {
+int64_t Lz4Codec::Compress(int64_t input_len, const uint8_t* input,
+    int64_t output_buffer_len, uint8_t* output_buffer) {
   return LZ4_compress(reinterpret_cast<const char*>(input),
       reinterpret_cast<char*>(output_buffer), input_len);
 }
